refactor(renderer): move welcome + activity views to self/home/
Step 5 of the renderer reorganization. personal-views/ is named for where
it renders (the non-group pane), not what it is about. Its home-landing
surfaces — the welcome view and the cross-group activity feed — are the
agent's own home, so they move to self/home/.

Pure file moves + import-path updates; no behaviour change.
