+++ src/renderer/src/self/home/activity-view/notification-asset.ts
@@ -9,17 +9,17 @@ import '@shoelace-style/shoelace/dist/components/icon/icon.js';
 import '@shoelace-style/shoelace/dist/components/button/button.js';
 import '@shoelace-style/shoelace/dist/components/dialog/dialog.js';
 
-import '../../applets/elements/applet-logo.js';
-import '../../applets/elements/applet-title.js';
-import '../../app/dialogs/loading-dialog.js';
-import { mossStoreContext } from '../../context.js';
+import '../../../applets/elements/applet-logo.js';
+import '../../../applets/elements/applet-title.js';
+import '../../../app/dialogs/loading-dialog.js';
+import { mossStoreContext } from '../../../context.js';
 import { consume } from '@lit/context';
-import { MossStore } from '../../moss-store.js';
+import { MossStore } from '../../../moss-store.js';
 import { AppletHash } from '@theweave/api';
 import { msg } from '@lit/localize';
 import { formatDistanceToNow } from 'date-fns';
-import { MossNotification } from '../../types.js';
-import { mossStyles } from '../../shared-styles.js';
+import { MossNotification } from '../../../types.js';
+import { mossStyles } from '../../../shared-styles.js';
 import { decodeHashFromBase64, AgentPubKey, DnaHashB64 } from '@holochain/client';
 
 @localized()
