+++ src/renderer/src/self/home/welcome-view/SnapTalkFunnies.json
